Check consensus hash consistency :
These changes use the hash of the consensus transaction set when
characterizing the mismatch between a locally built ledger and fully
validated network ledger. This allows detection of non-determinism in
transaction process, in which consensus succeeded, but a node somehow
generated a different subsequent ledger.

+++ src/test/app/CCLValidations_test.cpp
@@ -35,8 +35,8 @@ public:
     {
         testcase("Change validation trusted status");
         PublicKey key = derivePublicKey(KeyType::ed25519, randomSecretKey());
-        auto v = std::make_shared<STValidation>(
-            uint256(), NetClock::time_point(), key, calcNodeID(key), true);
+        auto v = std::make_shared<STValidation>(uint256(), uint256(),
+            NetClock::time_point(), key, calcNodeID(key), true);
 
         BEAST_EXPECT(!v->isTrusted());
         v->setTrusted();
